+++ GUI/DisasmFrame.cpp
@@ -5,94 +5,6 @@ wxBEGIN_EVENT_TABLE(DisasmFrame, wxFrame)
 	EVT_BUTTON(ButtonID::Step, DisasmFrame::OnStep)
 wxEND_EVENT_TABLE()
 
-
-DisasmListView::DisasmListView(std::shared_ptr<Debugger> d, wxWindow* parent)
-	: wxListView(parent, wxID_ANY, wxDefaultPosition, wxDefaultSize,
-		wxLC_REPORT | wxLC_VIRTUAL | wxLC_NO_HEADER),
-	debugger(d)
-{
-	wxListItem itemCol;
-	// Empty
-	itemCol.SetId(wxID_ANY);
-	itemCol.SetWidth(20);
-	InsertColumn(static_cast<long>(ColumnID::Empty), itemCol);
-	// Address
-	itemCol.SetId(wxID_ANY);
-	itemCol.SetWidth(45);
-	InsertColumn(static_cast<long>(ColumnID::Address), itemCol);
-	// Opcode
-	itemCol.SetId(wxID_ANY);
-	itemCol.SetWidth(70);
-	InsertColumn(static_cast<long>(ColumnID::Opcode), itemCol);
-	// Mnemonic
-	itemCol.SetId(wxID_ANY);
-	itemCol.SetWidth(70);
-	InsertColumn(static_cast<long>(ColumnID::Mnemonic), itemCol);
-	// Operand
-	itemCol.SetId(wxID_ANY);
-	itemCol.SetWidth(150);
-	InsertColumn(static_cast<long>(ColumnID::Operand), itemCol);
-	// Comment
-	itemCol.SetId(wxID_ANY);
-	itemCol.SetWidth(150);
-	InsertColumn(static_cast<long>(ColumnID::Comment), itemCol);
-
-	RefreshValues();
-}
-
-void DisasmListView::RefreshValues()
-{
-	std::shared_ptr<Disassembly> disasm(debugger->GetDisassembly());
-	addressTable = debugger->GetAddressTable();
-
-	disasmData.clear();
-
-	for (auto i : *disasm)
-	{
-		InsData data;
-		data.address = i.addressToStr();
-		data.opcode = i.opcodeToStr();
-		data.mnemonic = i.mnemonicToStr();
-		data.operand = i.operandsToStr();
-		data.comment = i.commentToStr();
-
-		disasmData.push_back(data);
-	}
-
-	SetItemCount(disasmData.size());
-	Refresh();
-
-	ShowAddress(debugger->cpuState.pc);
-}
-
-void DisasmListView::ShowAddress(uint16_t a)
-{
-	// Deselect previous item
-	Select(selectedItem, false);
-	// Select new item
-	selectedItem = addressTable[a];
-	Focus(selectedItem);
-	Select(selectedItem, true);
-}
-
-wxString DisasmListView::OnGetItemText(long item, long column) const
-{
-	if (static_cast<size_t>(item) > disasmData.size())
-	{
-		return "";
-	}
-
-	switch (static_cast<ColumnID>(column))
-	{
-	case ColumnID::Address:		return disasmData[item].address;
-	case ColumnID::Opcode:		return disasmData[item].opcode;
-	case ColumnID::Mnemonic:	return disasmData[item].mnemonic;
-	case ColumnID::Operand:		return disasmData[item].operand;
-	case ColumnID::Comment:		return disasmData[item].comment;
-	default: return "";
-	}
-}
-
 DisasmFrame::DisasmFrame(std::shared_ptr<Debugger> d, wxWindow* parent)
 	: wxFrame(parent, wxID_ANY, "DameDis", wxDefaultPosition, wxSize(600, 400)),
 	debugger(d)
